refactor(cms): split project query
Move the public project D1 reader and fallback mapping into cms/projects.ts so cms/index.ts stays focused on editor snapshot orchestration and public re-exports.

Verified with pnpm --filter @anipotts/lib typecheck, pnpm --filter @anipotts/lib test, pnpm --filter @anipotts/lib build, pnpm exec prettier --check packages/lib/src/cms/index.ts packages/lib/src/cms/projects.ts, git diff --check -- packages/lib/src/cms/index.ts packages/lib/src/cms/projects.ts, and pnpm dlx knip --no-exit-code.
