+++ packages/lib/src/cms/index.ts
@@ -6,15 +6,7 @@ import type {
   CmsWritingContent,
   NewsletterContent,
   PageContent,
-  Project,
-  ProjectCategory,
 } from "@anipotts/types";
-import { projectRowToProject } from "@anipotts/types";
-import { eq, asc, and } from "drizzle-orm";
-import { logger } from "../logger";
-import { projects as FALLBACK_PROJECTS } from "../data/projects";
-import { getDrizzle, parseJsonArray } from "../db";
-import * as s from "../db/schema";
 import { DEFAULT_CMS_PROJECTS, DEFAULT_CMS_WRITING } from "./defaults";
 import {
   cmsProjectPageKey,
@@ -50,6 +42,7 @@ export {
   fetchSiteSetting,
   fetchSocialLinks,
 } from "./settings";
+export { fetchProjects } from "./projects";
 export { fetchWriting, searchWriting } from "./writing";
 
 export async function fetchHomepageContent(): Promise<HomepageContent> {
@@ -107,74 +100,6 @@ export async function fetchCmsEditorSnapshot(): Promise<CmsEditorSnapshot> {
     newsletter: normalizeNewsletterContent(newsletterPage?.content),
     newsletterMeta: pageMeta(newsletterPage as PageContent<unknown> | null),
   };
-}
-
-// ---------------------------------------------------------------------------
-// Projects
-// ---------------------------------------------------------------------------
-
-export async function fetchProjects(options?: {
-  featured?: boolean;
-  category?: ProjectCategory;
-  visible?: boolean;
-  limit?: number;
-}): Promise<Project[]> {
-  const db = getDrizzle();
-  if (db) {
-    try {
-      const conditions = [];
-      const visibleFilter = options?.visible ?? true;
-      if (visibleFilter) {
-        conditions.push(eq(s.projects.visible, true));
-      }
-      if (options?.featured !== undefined) {
-        conditions.push(eq(s.projects.featured, options.featured));
-      }
-      if (options?.category) {
-        conditions.push(eq(s.projects.category, options.category));
-      }
-
-      const whereClause =
-        conditions.length > 0 ? and(...conditions) : undefined;
-
-      const results = await db
-        .select()
-        .from(s.projects)
-        .where(whereClause)
-        .orderBy(asc(s.projects.sort_order))
-        .limit(options?.limit ?? 1000);
-
-      if (results.length === 0) return FALLBACK_PROJECTS;
-      return results.map((row) =>
-        projectRowToProject({
-          ...row,
-          subtitle: row.subtitle ?? "",
-          description: row.description ?? "",
-          year: row.year ?? "",
-          category: (row.category ?? "project") as Project["category"],
-          role: row.role ?? "",
-          duration: row.duration ?? "",
-          tags: parseJsonArray(row.tags),
-          status: (row.status ?? "live") as
-            | "live"
-            | "in-progress"
-            | "coming-soon",
-          featured: row.featured ?? false,
-          visible: row.visible ?? true,
-          sort_order: row.sort_order ?? 0,
-          created_at: row.created_at ?? "",
-          updated_at: row.updated_at ?? "",
-        }),
-      );
-    } catch (err) {
-      logger.warn("cms", "D1 fetchProjects() failed, using fallback", {
-        error: String(err),
-      });
-      return FALLBACK_PROJECTS;
-    }
-  }
-
-  return FALLBACK_PROJECTS;
 }
 
 export type { ThoughtSummary, WritingSummary } from "@anipotts/types";
