+++ packages/lib/src/cms/projects.ts
@@ -1,0 +1,71 @@
+import type { Project, ProjectCategory } from "@anipotts/types";
+import { projectRowToProject } from "@anipotts/types";
+import { and, asc, eq } from "drizzle-orm";
+import { projects as FALLBACK_PROJECTS } from "../data/projects";
+import { getDrizzle, parseJsonArray } from "../db";
+import * as s from "../db/schema";
+import { logger } from "../logger";
+
+export async function fetchProjects(options?: {
+  featured?: boolean;
+  category?: ProjectCategory;
+  visible?: boolean;
+  limit?: number;
+}): Promise<Project[]> {
+  const db = getDrizzle();
+  if (db) {
+    try {
+      const conditions = [];
+      const visibleFilter = options?.visible ?? true;
+      if (visibleFilter) {
+        conditions.push(eq(s.projects.visible, true));
+      }
+      if (options?.featured !== undefined) {
+        conditions.push(eq(s.projects.featured, options.featured));
+      }
+      if (options?.category) {
+        conditions.push(eq(s.projects.category, options.category));
+      }
+
+      const whereClause =
+        conditions.length > 0 ? and(...conditions) : undefined;
+
+      const results = await db
+        .select()
+        .from(s.projects)
+        .where(whereClause)
+        .orderBy(asc(s.projects.sort_order))
+        .limit(options?.limit ?? 1000);
+
+      if (results.length === 0) return FALLBACK_PROJECTS;
+      return results.map((row) =>
+        projectRowToProject({
+          ...row,
+          subtitle: row.subtitle ?? "",
+          description: row.description ?? "",
+          year: row.year ?? "",
+          category: (row.category ?? "project") as Project["category"],
+          role: row.role ?? "",
+          duration: row.duration ?? "",
+          tags: parseJsonArray(row.tags),
+          status: (row.status ?? "live") as
+            | "live"
+            | "in-progress"
+            | "coming-soon",
+          featured: row.featured ?? false,
+          visible: row.visible ?? true,
+          sort_order: row.sort_order ?? 0,
+          created_at: row.created_at ?? "",
+          updated_at: row.updated_at ?? "",
+        }),
+      );
+    } catch (err) {
+      logger.warn("cms", "D1 fetchProjects() failed, using fallback", {
+        error: String(err),
+      });
+      return FALLBACK_PROJECTS;
+    }
+  }
+
+  return FALLBACK_PROJECTS;
+}
